+++ app/src/main/java/com/wavesplatform/wallet/v2/data/database/TransactionSaver.kt
@@ -1,4 +1,4 @@
-package com.wavesplatform.wallet.v2.database
+package com.wavesplatform.wallet.v2.data.database
 
 import com.vicpin.krealmextensions.*
 import com.wavesplatform.wallet.v2.data.Constants
