+++ app/src/main/java/com/wavesplatform/wallet/v2/data/helpers/AuthHelper.kt
@@ -5,8 +5,8 @@ import com.vicpin.krealmextensions.saveAll
 import com.wavesplatform.wallet.v1.util.PrefsUtil
 import com.wavesplatform.wallet.v2.data.Constants
 import com.wavesplatform.wallet.v2.data.model.remote.response.*
-import com.wavesplatform.wallet.v2.database.DBHelper
-import com.wavesplatform.wallet.v2.database.RealmMigrations
+import com.wavesplatform.wallet.v2.data.database.DBHelper
+import com.wavesplatform.wallet.v2.data.database.RealmMigrations
 import com.wavesplatform.wallet.v2.ui.home.profile.address_book.AddressBookUser
 import com.wavesplatform.wallet.v2.util.MigrationUtil
 import io.realm.Realm
